removed verbs from api
