+++ pages/api/pets/[uuid].ts
