+++ pages/api/plants/[uuid].ts
