+++ pages/house.tsx
@@ -1,7 +1,33 @@
-import React from 'react';
+import { Text } from '@chakra-ui/react';
+import React, { useEffect } from 'react';
 
 const House = () => {
-  return <div>House</div>;
+  const [house, setHouse] = React.useState<any>();
+
+  useEffect(() => {
+    async function getHouse(uuid: string) {
+      const res = await fetch(`/api/house/${uuid}`);
+      const data = await res.json();
+
+      return data;
+    }
+    const house = getHouse('123');
+    house.then((data) => setHouse(data));
+
+    return () => {};
+  }, []);
+
+  return (
+    <div>
+      {house && (
+        <>
+          {' '}
+          <Text>{house.emergencyContact?.name}</Text>
+          <Text>{house.address?.street}</Text>
+        </>
+      )}
+    </div>
+  );
 };
 
 export default House;
